# Conflicts:
#	cli/bootnode/boot_node.go
#	network/discovery/dv5_service.go
#	network/p2p/p2p_setup.go
#	utils/boot_node/node.go

+++ doppelganger/doppelganger.go
@@ -196,7 +196,7 @@ func (h *handler) Start(ctx context.Context) error {
 			// Perform liveness checks during the first run or at the last slot of the epoch.
 			// This ensures that the beacon node has had enough time to observe blocks and attestations,
 			// preventing delays in marking a validator as safe.
-			if (!firstRun && currentSlot%slotsPerEpoch != slotsPerEpoch-1) || startEpoch == currentEpoch {
+			if (!firstRun && uint64(currentSlot)%slotsPerEpoch != slotsPerEpoch-1) || startEpoch == currentEpoch {
 				continue
 			}
